fix(deps): break metamask.js and actions.ts cycle via metamask-state-basic
- Add ui/selectors/metamask-state-basic.js with getAlertEnabledness, getUnconnectedAccountAlertEnabledness, getCompletedOnboarding
- actions.ts imports these from metamask-state-basic instead of ducks/metamask
- metamask.js re-exports from metamask-state-basic and removes duplicate definitions
- Cycle count: 8 -> 7

Made-with: Cursor
